split xfile.h
